Add password cracking tools: Hydra and John integration
