+++ mcp_tools/password_cracking/hydra.py
@@ -1,0 +1,46 @@
+# mcp_tools/password_cracking/hydra.py
+
+from typing import Dict, Any
+
+def register_hydra_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def hydra_attack(
+        target: str,
+        service: str,
+        username: str = "",
+        username_file: str = "",
+        password: str = "",
+        password_file: str = "",
+        additional_args: str = ""
+    ) -> Dict[str, Any]:
+        """
+        Execute Hydra for password brute forcing with enhanced logging.
+
+        Args:
+            target: The target IP or hostname
+            service: The service to attack (ssh, ftp, http, etc.)
+            username: Single username to test
+            username_file: File containing usernames
+            password: Single password to test
+            password_file: File containing passwords
+            additional_args: Additional Hydra arguments
+
+        Returns:
+            Brute force attack results
+        """
+        data = {
+            "target": target,
+            "service": service,
+            "username": username,
+            "username_file": username_file,
+            "password": password,
+            "password_file": password_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔑 Starting Hydra attack: {target}:{service}")
+        result = hexstrike_client.safe_post("api/tools/hydra", data)
+        if result.get("success"):
+            logger.info(f"✅ Hydra attack completed for {target}")
+        else:
+            logger.error(f"❌ Hydra attack failed for {target}")
+        return result
