+++ mcp_tools/web_app_security/additional_security_tools.py
@@ -5,81 +5,6 @@ from typing import Dict, Any
 def register_additional_security_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def hydra_attack(
-        target: str,
-        service: str,
-        username: str = "",
-        username_file: str = "",
-        password: str = "",
-        password_file: str = "",
-        additional_args: str = ""
-    ) -> Dict[str, Any]:
-        """
-        Execute Hydra for password brute forcing with enhanced logging.
-
-        Args:
-            target: The target IP or hostname
-            service: The service to attack (ssh, ftp, http, etc.)
-            username: Single username to test
-            username_file: File containing usernames
-            password: Single password to test
-            password_file: File containing passwords
-            additional_args: Additional Hydra arguments
-
-        Returns:
-            Brute force attack results
-        """
-        data = {
-            "target": target,
-            "service": service,
-            "username": username,
-            "username_file": username_file,
-            "password": password,
-            "password_file": password_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔑 Starting Hydra attack: {target}:{service}")
-        result = hexstrike_client.safe_post("api/tools/hydra", data)
-        if result.get("success"):
-            logger.info(f"✅ Hydra attack completed for {target}")
-        else:
-            logger.error(f"❌ Hydra attack failed for {target}")
-        return result
-
-    @mcp.tool()
-    def john_crack(
-        hash_file: str,
-        wordlist: str = "/usr/share/wordlists/rockyou.txt",
-        format_type: str = "",
-        additional_args: str = ""
-    ) -> Dict[str, Any]:
-        """
-        Execute John the Ripper for password cracking with enhanced logging.
-
-        Args:
-            hash_file: File containing password hashes
-            wordlist: Wordlist file to use
-            format_type: Hash format type
-            additional_args: Additional John arguments
-
-        Returns:
-            Password cracking results
-        """
-        data = {
-            "hash_file": hash_file,
-            "wordlist": wordlist,
-            "format": format_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔐 Starting John the Ripper: {hash_file}")
-        result = hexstrike_client.safe_post("api/tools/john", data)
-        if result.get("success"):
-            logger.info(f"✅ John the Ripper completed")
-        else:
-            logger.error(f"❌ John the Ripper failed")
-        return result
-
-    @mcp.tool()
     def wpscan_analyze(url: str, additional_args: str = "") -> Dict[str, Any]:
         """
         Execute WPScan for WordPress vulnerability scanning with enhanced logging.
